Adds SVM Classifier functionality
Added Hinge loss function as an option for the SGDClassifier. Also added
support for non-numeric classes in training set. Logistic regression
might be a little buggy and should be reviewed. It gives decent results,
but not as good as SciKit Learn. Whereas the SVM gives the exact same
results as SciKit Learn. Tested with abalone data found here:

https://archive.ics.uci.edu/ml/datasets/Abalone

NOTE: Performance is iffy, trains in sub 10s, but it may be possible to
improve it. Possible issue is that our stop condition isn't good enough
so too many epochs are being run. May want to use ADAGRAD adaptive
learning rate to help with convergence.